[web] introduce community header
Summary:
This diff puts all the community header components I built in previous diffs and introduces it as the community header component. Just also want to note that this diff also includes some "hacky" logic I introduced to test the two main states of this header. This logic will be removed once I start working on improving the navigation state when navigating communites:

1. When home/all communities is selected (there should be no community action buttons and the header should say "All communities")
{F1013715}

2. When a community is selected (the name/avatar of the community should be in the header and the community action buttons should be rendered)
{F1013716}

Linear task: https://linear.app/comm/issue/ENG-5965/top-bar-when-a-community-is-selected and https://linear.app/comm/issue/ENG-5964/top-bar-when-inboxhome-is-selected

Depends on D10498

Test Plan:
Please see the screenshots below

When home/all communities is selected:
{F1013718}

When a community is selected:
{F1013719}

Reviewers: atul, kamil, rohan

Reviewed By: atul

Subscribers: varun, ashoat, tomek

Differential Revision: https://phab.comm.dev/D10499