[web] introduce panel component
Summary:
This component introduces the "Panel" component. I built the "Panel" component as the new UI is moslty made up of these "Panels" to house various parts of Comm, and I wanted to make sure we could easily enforce the style rules that every "Panel" will need. The "Panel" component is made up of two parts: the header and the body.

Here are some examples of the header:
{F1013892}

{F1013891}

{F1013890}

{F1013889}

Here are examples of the body:
{F1013900}

{F1013901}

{F1013902}

{F1013903}

Another thing the "Panel" component needs to consider is that a "Panel" component can have multiple sections. For example, in the screenshot below the app list has one section but the chat card has two sections (the thread list, and the message list). In the screenshot you can see that the message list has a darker background shade than the thread list, indicating that there are two sections here.
{F1013886}

Here is an example of a card with three sections (thread list, message list, and members list). Please note that the thread list and members list have the same lighter shade of black compared to the message list)
{F1013887}

Depends on 10499

Test Plan:
Please see the following screenshots

Here is an example of single section card (should be no visual differences which is expected)
{F1013906}

Here is an example of a two section card:
{F1013911}

Here is an example of a three section card:
{F1013968}

safari (checking for the shadow styles):
{F1023677}

firefox (checking for the shadow styles):
{F1023678}

Reviewers: atul, kamil, rohan

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10500