[web] upate emoji avatar selection modal to use new tabs component
Summary:
This diff updates the emoji avatar selection modal to use the new tab components. Since we decoupled the tabs from the tab content with this new tab component, the emoji avatar selection modal now needs to be responsible for rendering the correct tab content.

Depends on D10516

Test Plan:
Please see the demo video below to confirm that there are no regressions

{F1026063}

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10518