[web] update members modal to use the new tabs component
Summary:
This diff updates the members modal to use the new tab components. Since we decoupled the tabs from the tab content with this new tab component, the members modal now needs to be responsible for rendering the correct tab content.

Depends on D10520

Test Plan:
Please see the demo video to confirm that there are no regressions

{F1026288}

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10521