[web] update chat tabs to use the new tabs component
Summary:
This diff updates the chat tabs to use the new tab components. Since we decoupled the tabs from the tab content with this new tab component, the chat tabs now needs to be responsible for rendering the correct tab content.

Depends on D10522

Test Plan:
Please see the demo video below to confirm that there are no regressions

{F1026408}

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10523