[web] Fix default apps
Summary:
I noticed that on web in default state we used `defaultEnabledApps` instead of `defaultWebEnabledApps`. This caused the calendar not to be visible to a just logged in user. This was also inconsitent with the fact that in the `reduceEnabledApps` reducer we set `enabledApps` to `defaultWebEnabledApps` on logout.
This was introduced in D9145 and Michał says that it's a mistake.

Test Plan: Tested that the correct default is used now.

Reviewers: michal, ginsu

Reviewed By: michal, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10525