[native] Upload and download logs
Summary:
Testing the backup log functionality on mobile devices. This is more concerned with the networking/websocket layer working and less with business logic. Business logic is tested in the previous diff, in `commtest`.

Depends on D10534

Test Plan:
Created and restored backup. Checked that there are no errors.

This was tested on iOS and Android emulators. I will also test this on physical devices before landing.

Reviewers: bartek, kamil

Subscribers: