[native] Update `DisconnectedBar` to display based on `state.connectivity.connected`
Summary:
Similar to D10469, but for `native`.

As mentioned in that diff, going to start by updating the `DisconnectedBar` on both `native` and `web` client to use network connectivity info instead of `keyserver` connectivity info.

Will then go ahead and rename eg `useDisconnectedBarVisibilityHandler` to make it clear that they're no longer connected to `DisconnectedBar`.

---

Depends on D10469

Test Plan:
1. Open `native` app on physical device
2. Kill local `keyserver`
3. Observe that nothing happens
4. Turn off WiFi
5. Observe `DisconnectedBar`

Reviewers: ashoat, ginsu, tomek

Reviewed By: ashoat, ginsu

Subscribers: inka

Differential Revision: https://phab.comm.dev/D10475