[RN72] Update Jest Flow libdefs
Summary:
@atul [called out](https://phab.comm.dev/D10536#304294) that these libdefs are outdated in D10536. This diff updates them all to the latest version.

I updated the libdefs by first running `git rm` on each of them, and then running `flow-typed install jest` in each of the directories.

I also noticed that `web` was missing the libdef, so I added it.

Not sure these libdefs are currently used, but better to keep them updated then out-of-date, especially if we ever introduce any usages in the future.

Depends on D10544

Test Plan: Ran `yarn flow-all` from the workspace root. Also CI will run on this diff

Reviewers: atul

Subscribers: