[web] rename tabs to tabs legacy
Summary:
Rename diff.

In order to house the thread list in the panel component, we need to refactor the tabs component. At this moment the tabs component is made of two parts: the tabs and the tab content. The tabs component utilizes the `React.Children.*` API to interact with those child components. The `React.Children.*` API is listed as Legacy React API in the react docs and additionally for the redesign I need to decouple the tabs from the tab content to put the tabs in the panel header and the tab content in the panel body.

In order to make this refactor easy to review, I broke this down into many steps. The first step is to rename this file and the imports to "tabs legacy" and then in a subsequent diff I will introduce a new tabs component which will utilize an alternate solution outlined in the react docs: https://react.dev/reference/react/Children#accepting-an-array-of-objects-as-a-prop

Afterwards I will update every tab use case to use the new tabs api

Depends on D10515

Test Plan: `flow` and ensured that all the tabs worked as expected

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10517