[web] factor out pinned messages banner into its own component
Summary: We will be using the thread top bar component as the panel header for the message list panel item. In order to accomplish this we will need to decouple the pinned messages banner from the top bar. To make the move a bit easier/more orgnaized, I took a second to make the pinned messages banner into its own component

Test Plan: `flow` and confirmed the pinned messages banner still works as expceted

Reviewers:

Subscribers: