[native] Skip onInitialAppLoad if usingCommServicesAccessToken
Summary:
This handles part of [ENG-5940](https://linear.app/comm/issue/ENG-5940/update-fetchnewcookiefromnativecredentials-callsites) (`appStartReduxLoggedInButInvalidCookie` and `appStartCookieLoggedInButInvalidRedux`).

In the old format, `currentUserInfo` and the `cookie` from my keyserver should always be in sync. In the multi-keyserver world, `currentUserInfo` is sourced from Comm services while `cookie` is keyserver-specific, so there is no issue when the two are out-of-sync.

If a keyserver is logged out when we have a CSAT, the keyserver connection handlers will handle it by logged the user into the keyserver using the Olm auth responder (assuming there isn't a permanent `connectionIssue`).

We'll have a somewhat similar issue with `currentUserInfo` and the CSAT potentially being out-of-sync. This will be handled separately; will make a comment on [ENG-5798](https://linear.app/comm/issue/ENG-5798/refactor-loggedoutmodal) to discuss this further, but it looks like it's meant to be part of [ENG-4332](https://linear.app/comm/issue/ENG-4332/get-logged-in-clients-to-registerlog-in-with-identity-service).

Test Plan: Not sure how to test this one. Flow checks out

Reviewers: inka!, varun, tomek

Differential Revision: https://phab.comm.dev/D10575