[lib][native] Rename fetchNewCookieFromNativeCredentials to resolveKeyserverSessionInvalidation
Summary:
In [ENG-5940](https://linear.app/comm/issue/ENG-5940/update-fetchnewcookiefromnativecredentials-callsites), I'm working on refactoring this function.

One of the callsites will be replaced with a different function that will handle whole-app login via identity service.

The rest of the callsites will be refactored. They will continue trying to automatically log in to the keyserver, but if `usingCommServicesAccessToken` is set, the automatic login will be performed via the Olm auth responder instead of by using native credentials.

To reflect the new functionality, I want to start by renaming the function.

Depends on D10575

Test Plan: Flow

Reviewers: inka!, varun, tomek

Subscribers: