[lib][native] Rename resolveInvalidatedCookie to resolveKeyserverSessionInvalidationUsingNativeCredentials
Summary:
Similar to the parent diff. As part of [ENG-5940](https://linear.app/comm/issue/ENG-5940/update-fetchnewcookiefromnativecredentials-callsites), we're going to be reducing our reliance on this function. If the client is `usingCommServicesAccessToken`, then instead of calling this function we'll just call the Olm auth responder.

As such, it's probably a good idea to rename the function to be more clear about what it does.

Depends on D10576

Test Plan: Flow

Reviewers: inka, varun, tomek

Reviewed By: inka

Differential Revision: https://phab.comm.dev/D10577