[lib] Introduce canResolveKeyserverSessionInvalidation
Summary:
Previously, to figure out if it was possible to resolve a keyserver session invalidation, we were checking for the presence of the `resolveInvalidatedCookie` function.

That function was missing on `web` but present on `native`, since it required native credentials to work.

Now we can attempt resolution for any client that is `usingCommServicesAccessToken`, so we want to update the check. Instead of duplicating the code, I introduced a function to handle the check.

Depends on D10577

Test Plan: Will test this by triggering a session invalidation by deleting a cookie from the MariaDB database

Reviewers: inka, varun, tomek

Reviewed By: inka

Differential Revision: https://phab.comm.dev/D10578