[web] introduce panel header component
Summary:
As I have been working on the redesign and have been converting our ui to utilize the new `Panel` component, I noticed that a lot of the headers followed a similar ui pattern. Rather than have this pattern be copied and pasted a bunch, I created this `PanelHeader` component that can be used to factor and abstract that logic out. The pattern I was noticing with the panel headers was that many panel headers either would just have the label, or they would have an optional add button.

{F1055842}

{F1055841}

{F1055840}

{F1055839}

{F1055838}

Test Plan: `flow` + confirmed that there were no regressions with the app list header

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10610