[lib][native][web] Expose identity API by using a context
Summary:
Create a context that exposes the API client and recreates it on every user change or token change. Simplify the client by avoiding a need to provide credentials on every call to the API.

Depends on D10449

https://linear.app/comm/issue/ENG-6404/move-identity-client-to-a-context

Test Plan: Checked if credentials are correctly provided to a client and if the proper context value is provided (on native and web).

Reviewers: inka, kamil, bartek, marcin, ashoat

Reviewed By: ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D10584