[lib] Get keyserver keys in a handler
Summary:
Get identity client from a context and use it to get the keys. This diff doesn't perform anything useful with them, but presents how they are received, and the code will be extended later in the stack.

https://linear.app/comm/issue/ENG-6084/call-the-getkeyserverkeys-in-the-handler

Depends on D10584

Test Plan: Add sole logs that check if the method is properly called.

Reviewers: inka, varun, kamil, bartek

Reviewed By: varun

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D10585