[web] update grpc-web codegen
Summary:
this update reflects the changes to InboundKeysForUserResponse made in D10560.

the codegen had a hard time with the oneOf's, so had to make some manual changes to the .flow file

Test Plan:
called getInboundKeysForUser on my local identity service from root.js and got back a response containing a wallet address and social proof.
called it again and successfully got a username + keys for a username/password user.

Reviewers: ashoat, bartek

Reviewed By: ashoat, bartek

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D10587