[backup] Allow one websocket connection to handle multiple backups
Summary:
After working on a final backup client for native I realized that the backup service log API would be better if it wasn't opened per backup, and instead the messages contained the information which backup they apply to. This allow us to to run only one websocket connection on the backup client and handle all edge cases (e.g. there was no network connection, but we started creating a new backup and now have logs for two backups). This simplifies the backup client code a lot because we don't need to spawn new webosocket connection for every backup_id and keep track of when to kill them. This is a breaking change but backup service isn't used in production app (there is only staff-only opt-in testing code) so it should be fine.

Test Plan:
Run integration tests.

Reviewers: bartek, kamil

Subscribers: