[lib] Fix types in deleteThread
Summary:
issue: [[ https://linear.app/comm/issue/ENG-6407/fix-delete-thread-calls | ENG-6407 ]]
When I refactored deleteThread in D9371, I forgot to change the type taken by it. This caused no problems, because flow can't really tell the difference between `CallServerEndpoint` and `CallKeyserverEndpoint`, so it didn't have a problem with that we were passing `CallKeyserverEndpoint` to it.
This is just a type change, the code worked all the time

Test Plan: Tested that deleting threads works

Reviewers: michal, kamil, ginsu

Reviewed By: michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10591