[web] make mouse pointer when hovering over checkbox input
Summary: Our new add/select users modal uses checkboxes to select a group of users. As I was working on building this modal I noticed that when I hovered over a checkbox the mouse was not a pointer. I thought that was a bit strange and updated the behavior so that the mouse would be a pointer whenever a checkbox is hovered over

Test Plan:
Please see the demo videos below

before:
{F1071331}

after:
{F1071332}

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10717