[web] fix prop drilling within relationships modals
Summary: As I was working with the relationship modals/panels I found that we were drilling this `onClose` prop method that pretty much was just `popModal`. We don't need to drill this down everywhere and instead we can just call `popModal` using `useModalContext` hook where we need it.

Test Plan: `flow` + confirmed that the modal on close behavoir was popping as expected

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10743