[native] Backup log upload
Summary:
[ENG-5560 : Upload logic for logs](https://linear.app/comm/issue/ENG-5560/upload-logic-for-logs)

Similarly to the previous diff we want to upload log files to the backup service. One difference is that we need to wait for upload confirmation from the backup service, sent as a WS message, and only then remove the backup file. To prevent sending the same log files multiple times we keep track of logs that were sent but not yet confirmed in a `HashSet` that is shared between sender and receiver futures.

Depends on D10632

Test Plan: Create `backup-123-log-456` and `backup-123-log-456-attachments` files. Enable backup upload. Make sure all data was successfuly uploaded to the backup service and the files were removed.

Reviewers: marcin, bartek, kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10633