[lib] First pass at removing `LegacyRawThreadInfo`
Summary:
Part of https://linear.app/comm/issue/ENG-6366/rename-minimallyencodedrawthreadinfo-to-rawthreadinfo

We want to reduce usages of `LegacyRawThreadInfo` as much as possible. It should really only be around in `keyserver`/`lib` to handle old clients and in `web`/`native` to handle old migrations.

This was the first pass of removing `LegacyRawThreadInfo` where it was easy. Will continue trying to rip out as much of `LegacyRawThreadInfo` as possible.

---

Depends on D10625

Test Plan:
`flow`

Reviewers:
ashoat, ginsu, tomek