[lib] Remove `LegacyThreadInfo` from `thread-selectors`
Summary:
Part of https://linear.app/comm/issue/ENG-6366/rename-minimallyencodedrawthreadinfo-to-rawthreadinfo

With the prior `flow` refactor for minimal encoded permissions, we had to start at the "leaf nodes" (hooks and components) and work back up. For this refactor we need to do the opposite. We've removed `Legacy*` types, now we're removing them from selectors, and then we'll work from react root component down to leaves.

---

Depends on D10643

Test Plan: `flow`

Reviewers: ashoat, ginsu, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D10644