[lib] Remove threads from calendar filters on logout
Summary:
issue: [[ https://linear.app/comm/issue/ENG-5988/remove-only-thread-ids-related-to-the-keyserver-that-was-logged-out-of | ENG-5988 ]]
We want to remove all ids of threads that belong to the keyserver the user is logging out of.
IT'S VERY IMPORTANT THAT THE CURRENT BEHAVIOUR STAYS AS IS FOR USERS NOT USING CSAT. It will be the case for all diffs in this stack. Logout / delete account actions will not be replaced like login was, so we are using the same action types and hiding the old logic behind the flag.

Test Plan: Tested that if the user logs out and `usingCommServicesAccessToken === true` then threads that belong to the keyserver the user is logging out of are deleted. Tested that if `usingCommServicesAccessToken === false`, then filters are set to default values

Reviewers: michal, kamil, ginsu

Reviewed By: michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10647