[lib] Refactor logout in reduceEntryInfos
Summary:
issue: [[ https://linear.app/comm/issue/ENG-5766/refactor-logout-delete-account-logic-in-reducers | ENG-5766 ]]
We don't want to remove `lastUserInteractionCalendar` for users using CSAT.
The rest of the reducer is already correct, because it checks new threads aand based on that removes entries. So if thread reducer removes threads related to the keyserver, the entires will be removed as well

Test Plan: Tested that if a user is connected to another keyserver and logs out of it, the corresponding entires are indeed removed. Tested that for a user not using CSAT, `lastUserInteractionCalendar` is still reset to 0

Reviewers: michal, kamil, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10650